requirements/base.txt
# Web / Server
fastapi==0.115.0
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-dotenv==1.0.1

# Data / Validation
pydantic==2.12.5
pydantic-settings==2.7.0
email-validator==2.2.0
tzdata==2024.2
python-multipart==0.0.18

# MongoDB (Motor 3.7.x supports pymongo 4.10+)
motor==3.7.1
pymongo==4.10.1

# Auth / Security
authlib==1.4.0
cryptography==43.0.3
bcrypt==4.2.1
argon2-cffi==23.1.0
passlib==1.7.4
pyjwt==2.10.1

# SQL Server / Excel / Imaging
pyodbc==5.2.0
openpyxl==3.1.5
xlsxwriter==3.2.0
Pillow==11.0.0

# HTTP / Retry / System
requests==2.32.3
tenacity==9.1.2
psutil==6.1.0

# Redis
redis==5.2.1
redis[hiredis]==5.2.1

# Data science utilities (keep here if used in runtime endpoints/jobs)
numpy==2.0.2
pandas==2.2.3

# Packaging (security bump)
setuptools==78.1.1
pypika==0.51.1
